@@ -3,5 +3,5 @@ source 'https://rubygems.org'
 gemspec
 
 group :test do
-  gem 'rails', '3.2.22'
+  gem 'rails', '4.2.7'
 end
