@@ -6,8 +6,7 @@ Dummy::Application.configure do
   config.action_mailer.raise_delivery_errors = false
   config.active_support.deprecation = :log
   config.action_dispatch.best_standards_support = :builtin
-  config.active_record.mass_assignment_sanitizer = :strict
-  config.active_record.auto_explain_threshold_in_seconds = 0.5
   config.assets.compress = false
   config.assets.debug = true
+  config.eager_load = false
 end
