@@ -1,6 +1,6 @@
 Dummy::Application.configure do
   config.cache_classes = true
-  config.serve_static_assets = true
+  config.serve_static_files = true
   config.static_cache_control = 'public, max-age=3600'
   config.whiny_nils = true
   config.consider_all_requests_local = true
@@ -8,7 +8,7 @@ Dummy::Application.configure do
   config.action_dispatch.show_exceptions = false
   config.action_controller.allow_forgery_protection = false
   config.action_mailer.delivery_method = :test
-  config.active_record.mass_assignment_sanitizer = :strict
   config.active_support.deprecation = :stderr
   config.authorization_debug = true
+  config.eager_load = false
 end
