@@ -1,3 +1,3 @@
 module Breadcrumbs
-  VERSION = '0.0.1'
+  VERSION = '0.0.1'.freeze
 end
