@@ -1,7 +1,4 @@
-require 'active_model/validations'
-require 'active_model/naming'
-require 'active_model/translation'
-require 'active_model/dirty'
+require 'active_model'
 
 require_relative 'custom_validators/email_validator'
 require_relative 'custom_validators/url_validator'
