@@ -10,11 +10,9 @@ Dummy::Application.configure do
 
   config.action_dispatch.best_standards_support = :builtin
 
-  config.active_record.mass_assignment_sanitizer = :strict
-
-  config.active_record.auto_explain_threshold_in_seconds = 0.5
-
   config.assets.compress = false
 
   config.assets.debug = true
+
+  config.eager_load = false
 end
