@@ -1,3 +1,3 @@
 module Presenter
-  VERSION = "0.0.1"
+  VERSION = '0.0.1'.freeze
 end
