@@ -1,6 +1,6 @@
 FactoryGirl.define do
   factory :restrictions_set, class: 'Restrictions::Set' do
-    label { Faker::Lorem.sentence }
+    label { FFaker::Lorem.sentence }
     sequence(:identifier) { |n| "identifier#{n}"}
 
     trait :pure do
